requirements.txt
anthropic>=0.30.0
openai>=1.30.0
google-genai>=0.8.0
python-docx>=1.1.0
rich>=13.7.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
japanize-matplotlib>=1.1.3
Pillow>=10.0.0
streamlit>=1.33.0
